requirements.txt
fastapi 
uvicorn
fastai
pillow
openai
python-dotenv
opencv-python>=4.8.0
python-multipart